+++ src/Page/task-list/index.tsx
@@ -1,9 +1,10 @@
 import * as React from "react";
 import { useSelector } from "react-redux";
 
-import { ColumnItem } from "../../Compoents/Column";
+import { ColumnItem } from "../../widgets";
+import { AddColumn } from "../../features";
 import { RootState } from "../../app/providers/with-redux";
-import { AddColumn } from "../../Compoents/Column/AddColumn";
+
 import { Layout } from "./styles";
 
 export const TaskListPage: React.FC = () => {
