+++ src/features/columns/Components/index.ts
@@ -1,0 +1,5 @@
+export * from "./TasksColumn";
+export * from "./AddColumn";
+export * from "./ColumnsMenu";
+export * from "./ComponentCreatingTasks";
+export * from "./HeaderColumn";
