+++ src/widgets/columns/components/index.tsx
@@ -1,9 +1,7 @@
 import React from "react";
-import { ContainerColumn } from "../../style/column";
-import { HeaderColumn } from "./HeaderColumn";
-import { TasksColumn } from "./TasksColumn";
-import { IColumnsSlices } from "../../state/slices/columnsSlices";
-import { ComponentCreatingTasks } from "./ComponentCreatingTasks";
+import { IColumnsSlices } from "../libs";
+import { ContainerColumn } from "../../../shared";
+import { ComponentCreatingTasks, HeaderColumn, TasksColumn } from "../../../features";
 
 
 export const ColumnItem: React.FC<IColumnsSlices> = (props) => {
