+++ src/widgets/columns/libs/types.ts
@@ -1,0 +1,10 @@
+export interface ITasks {
+  id: string,
+  name: string
+}
+
+export interface IColumnsSlices {
+  id: string;
+  name: string;
+  tasks: Array<ITasks>
+}
